Rename files and references to match project
