+++ src/awssh/__main__.py
@@ -1,5 +1,5 @@
 """Code to run if this package is used as a Python module."""
 
-from .ssm_ssh import main
+from .awssh import main
 
 main()
